Created groupby dialog from context menu
Right click on a column in the variant view to see a list
of values taken by the field in question
